Move DataBox related headers to src/DataStructures/DataBox dir
