+++ src/Evolution/Systems/GeneralizedHarmonic/Tags.hpp
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "DataStructures/DataBoxTag.hpp"
+#include "DataStructures/DataBox/DataBoxTag.hpp"
 
 class DataVector;
