+++ src/Executables/Benchmark/Benchmark.cpp
@@ -4,7 +4,7 @@
 #include <benchmark/benchmark.h>
 #include <vector>
 
-#include "DataStructures/DataBoxTag.hpp"
+#include "DataStructures/DataBox/DataBoxTag.hpp"
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/Index.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
