+++ src/Time/Actions/UpdateU.hpp
@@ -9,7 +9,7 @@
 #include <tuple>
 #include <utility>
 
-#include "DataStructures/DataBox.hpp"
+#include "DataStructures/DataBox/DataBox.hpp"
 #include "Parallel/ConstGlobalCache.hpp"
 #include "Time/Tags.hpp"
 #include "Time/Time.hpp"
