+++ tests/Unit/Evolution/Systems/ScalarWave/Test_Equations.cpp
@@ -3,7 +3,7 @@
 
 #include <memory>
 
-#include "DataStructures/DataBox.hpp"
+#include "DataStructures/DataBox/DataBox.hpp"
 #include "DataStructures/DataBox/Prefixes.hpp"
 #include "DataStructures/Index.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
