+++ tests/Unit/Parallel/Test_TypeTraits.cpp
@@ -5,7 +5,7 @@
 #include "AlgorithmGroup.hpp"
 #include "AlgorithmNodegroup.hpp"
 #include "AlgorithmSingleton.hpp"
-#include "DataStructures/DataBox.hpp"
+#include "DataStructures/DataBox/DataBox.hpp"
 #include "Parallel/TypeTraits.hpp"
 
 namespace {
