+++ tests/Unit/PointwiseFunctions/MathFunctions/Test_TensorProduct.cpp
@@ -6,7 +6,7 @@
 #include <memory>
 #include <utility>
 
-#include "DataStructures/DataBoxTag.hpp"
+#include "DataStructures/DataBox/DataBoxTag.hpp"
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/Index.hpp"
 #include "DataStructures/MakeWithValue.hpp"
